+++ samples/extensions/sparse_image/sparse_image.h
@@ -158,7 +158,7 @@ class SparseImage : public ApiVulkanSample
 
 		uint32_t get_size()
 		{
-			return memory_sectors.size();
+			return static_cast<uint32_t>(memory_sectors.size());
 		}
 
 		std::list<std::weak_ptr<MemSector>> &get_memory_sectors()
@@ -191,7 +191,7 @@ class SparseImage : public ApiVulkanSample
 
 			for (size_t i = 0U; i < pages_per_allocation; i++)
 			{
-				available_offsets.insert(page_size * i);
+				available_offsets.insert(static_cast<uint32_t>(page_size * i));
 			}
 		}
